requirements.txt
pytest==7.3.0
selenium==4.9.1
pytest-xdist==3.2.1
pytest-rerunfailures==11.1.2
pytest-repeat==0.9.1
allure-pytest==2.13.2
webdriver-manager==3.8.6
flake8==6.0.0
requests~=2.30.0
